Shortcut the extract task if hash match the done file

+++ samples-dev/src/templates/provisionable-tool-chains-plugin/src/main/groovy/org/gradle/samples/tasks/ExtractToolChainTask.groovy
@@ -1,6 +1,8 @@
 package org.gradle.samples.tasks
 
+import groovy.io.FileType
 import groovy.transform.CompileStatic
+import org.apache.commons.codec.digest.DigestUtils
 import org.apache.commons.compress.archivers.ArchiveEntry
 import org.apache.commons.compress.archivers.ArchiveInputStream
 import org.apache.commons.compress.archivers.ArchiveStreamFactory
@@ -20,7 +22,8 @@ import org.gradle.api.tasks.TaskAction
 import java.nio.file.Files
 import java.nio.file.Paths
 import java.nio.file.attribute.FileTime
-import java.nio.file.attribute.PosixFilePermission;
+import java.nio.file.attribute.PosixFilePermission
+import java.security.MessageDigest;
 
 @CompileStatic
 class ExtractToolChainTask extends DefaultTask {
@@ -37,57 +40,59 @@ class ExtractToolChainTask extends DefaultTask {
 
     @TaskAction
     private void extract() {
-        doneFile.delete()
-        FileUtils.deleteDirectory(toolChainLocation.get().asFile)
+        if (!isExtractUpToDate()) {
+            doneFile.delete()
+            FileUtils.deleteDirectory(toolChainLocation.get().asFile)
 
-        InputStream fileStream = new BufferedInputStream(Files.newInputStream(toolChainArchive.get().asFile.toPath()))
-        ArchiveInputStream archiveStream = createArchiveStream(new BufferedInputStream(createDecompressStream(fileStream)))
+            InputStream fileStream = new BufferedInputStream(Files.newInputStream(toolChainArchive.get().asFile.toPath()))
+            ArchiveInputStream archiveStream = createArchiveStream(new BufferedInputStream(createDecompressStream(fileStream)))
 
-        File outputDir = toolChainLocation.asFile.get()
-        try {
-            for (ArchiveEntry entry = archiveStream.nextEntry; entry != null; entry = archiveStream.nextEntry) {
-                String entryPath = entry.getName().split("[\\/]").drop(1).join("/")
-                final File file = new File(outputDir, entryPath)
-                if (entry.isDirectory()) {
-                    continue
-                }
-
-                file.getParentFile().mkdirs()
-                if (entry instanceof TarArchiveEntry && entry.isSymbolicLink()) {
-                    Files.createSymbolicLink(file.toPath(), Paths.get(entry.getLinkName()))
-                } else if (entry instanceof TarArchiveEntry && !entry.getLinkName().isEmpty()) {
-                    OutputStream outStream = file.newOutputStream()
-                    String sourcePath = entry.getLinkName().split("[\\/]").drop(1).join("/")
-                    File sourceFile = new File(outputDir, sourcePath)
-                    InputStream inStream = sourceFile.newInputStream()
-                    IOUtils.copy(inStream, outStream)
-                    IOUtils.closeQuietly(inStream)
-                    IOUtils.closeQuietly(outStream)
-
-                    Files.setLastModifiedTime(file.toPath(), FileTime.from(entry.getLastModifiedDate().toInstant()))
-
-                    if (entry instanceof TarArchiveEntry) {
-                        file.setExecutable((entry.getMode() & toModeBit(PosixFilePermission.OWNER_EXECUTE)) > 0, true)
+            File outputDir = toolChainLocation.asFile.get()
+            try {
+                for (ArchiveEntry entry = archiveStream.nextEntry; entry != null; entry = archiveStream.nextEntry) {
+                    String entryPath = entry.getName().split("[\\/]").drop(1).join("/")
+                    final File file = new File(outputDir, entryPath)
+                    if (entry.isDirectory()) {
+                        continue
                     }
-                } else {
-                    OutputStream outStream = file.newOutputStream()
-                    IOUtils.copy(archiveStream, outStream)
-                    IOUtils.closeQuietly(outStream)
 
-                    Files.setLastModifiedTime(file.toPath(), FileTime.from(entry.getLastModifiedDate().toInstant()))
+                    file.getParentFile().mkdirs()
+                    if (entry instanceof TarArchiveEntry && entry.isSymbolicLink()) {
+                        Files.createSymbolicLink(file.toPath(), Paths.get(entry.getLinkName()))
+                    } else if (entry instanceof TarArchiveEntry && !entry.getLinkName().isEmpty()) {
+                        OutputStream outStream = file.newOutputStream()
+                        String sourcePath = entry.getLinkName().split("[\\/]").drop(1).join("/")
+                        File sourceFile = new File(outputDir, sourcePath)
+                        InputStream inStream = sourceFile.newInputStream()
+                        IOUtils.copy(inStream, outStream)
+                        IOUtils.closeQuietly(inStream)
+                        IOUtils.closeQuietly(outStream)
 
-                    if (entry instanceof TarArchiveEntry) {
-                        file.setExecutable((entry.getMode() & toModeBit(PosixFilePermission.OWNER_EXECUTE)) > 0, true)
+                        Files.setLastModifiedTime(file.toPath(), FileTime.from(entry.getLastModifiedDate().toInstant()))
+
+                        if (entry instanceof TarArchiveEntry) {
+                            file.setExecutable((entry.getMode() & toModeBit(PosixFilePermission.OWNER_EXECUTE)) > 0, true)
+                        }
+                    } else {
+                        OutputStream outStream = file.newOutputStream()
+                        IOUtils.copy(archiveStream, outStream)
+                        IOUtils.closeQuietly(outStream)
+
+                        Files.setLastModifiedTime(file.toPath(), FileTime.from(entry.getLastModifiedDate().toInstant()))
+
+                        if (entry instanceof TarArchiveEntry) {
+                            file.setExecutable((entry.getMode() & toModeBit(PosixFilePermission.OWNER_EXECUTE)) > 0, true)
+                        }
                     }
                 }
+            } catch (EOFException ex) {
+                // Ignore EOF exception
+            } finally {
+                IOUtils.closeQuietly(archiveStream)
             }
-        } catch (EOFException ex) {
-            // Ignore EOF exception
+
+            doneFile.text = md5Hash(toolChainLocationFile)
         }
-
-        IOUtils.closeQuietly(archiveStream)
-
-        doneFile.createNewFile()
     }
 
     private static CompressorInputStream createDecompressStream(InputStream inputStream) {
@@ -100,5 +105,21 @@ class ExtractToolChainTask extends DefaultTask {
 
     private static int toModeBit(PosixFilePermission permission) {
         return 1 << permission.ordinal()
+    }
+
+    private boolean isExtractUpToDate() {
+        return toolChainLocationFile.exists() && doneFile.exists() && doneFile.text == md5Hash(toolChainLocationFile)
+    }
+
+    private File getToolChainLocationFile() {
+        return toolChainLocation.get().asFile
+    }
+
+    private static String md5Hash(File dirToHash) {
+        MessageDigest digest = DigestUtils.getDigest("MD5")
+        dirToHash.eachFileRecurse(FileType.FILES) { File fileToHash ->
+            digest.update(fileToHash.bytes)
+        }
+        return new String(digest.digest())
     }
 }
